requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
PyYAML
requests
python-dotenv